# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EnviroIQ ESG Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/passkey/register/begin:
    post:
      operationId: beginPasskeyRegistration
      tags: [auth]
      summary: Begin passkey registration (WebAuthn)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeginPasskeyRegistrationRequest"
      responses:
        "200":
          description: Registration options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyRegistrationOptions"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/passkey/register/complete:
    post:
      operationId: completePasskeyRegistration
      tags: [auth]
      summary: Complete passkey registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePasskeyRegistrationRequest"
      responses:
        "200":
          description: Registration complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "400":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/passkey/authenticate/begin:
    post:
      operationId: beginPasskeyAuthentication
      tags: [auth]
      summary: Begin passkey authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeginPasskeyAuthRequest"
      responses:
        "200":
          description: Authentication options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyAuthOptions"
  /auth/passkey/authenticate/complete:
    post:
      operationId: completePasskeyAuthentication
      tags: [auth]
      summary: Complete passkey authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePasskeyAuthRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/magic-link/request:
    post:
      operationId: requestMagicLink
      tags: [auth]
      summary: Request magic link email fallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/magic-link/verify:
    post:
      operationId: verifyMagicLink
      tags: [auth]
      summary: Verify magic link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkVerifyRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ────────────────────────────────────────────────────────────────
    SessionInfo:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, org_admin, org_viewer]
        organisationId:
          type: string
        organisationName:
          type: string
        isAuthenticated:
          type: boolean
      required:
        - userId
        - email
        - role
        - isAuthenticated
    BeginPasskeyRegistrationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
      required:
        - email
        - name
    PasskeyRegistrationOptions:
      type: object
      properties:
        challenge:
          type: string
        rp:
          type: object
          properties:
            name:
              type: string
            id:
              type: string
          required: [name, id]
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            displayName:
              type: string
          required: [id, name, displayName]
        pubKeyCredParams:
          type: array
          items:
            type: object
        timeout:
          type: integer
        attestation:
          type: string
      required:
        - challenge
        - rp
        - user
        - pubKeyCredParams
    CompletePasskeyRegistrationRequest:
      type: object
      properties:
        credential:
          type: object
          additionalProperties: true
        email:
          type: string
        name:
          type: string
      required:
        - credential
        - email
    BeginPasskeyAuthRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: []
    PasskeyAuthOptions:
      type: object
      properties:
        challenge:
          type: string
        rpId:
          type: string
        allowCredentials:
          type: array
          items:
            type: object
        timeout:
          type: integer
      required:
        - challenge
    CompletePasskeyAuthRequest:
      type: object
      properties:
        credential:
          type: object
          additionalProperties: true
      required:
        - credential
    MagicLinkRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    MagicLinkVerifyRequest:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    # ─── ORGANISATION ────────────────────────────────────────────────────────
...