# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EnviroIQ ESG Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInfoRole } from "./sessionInfoRole";

export interface SessionInfo {
  userId: string;
  email: string;
  name?: string;
  role: SessionInfoRole;
  organisationId?: string;
  organisationName?: string;
  isAuthenticated: boolean;
}